refactor: extract mock data from useApplicationConformaResults
Move `generateMockResults()` (~190 lines of dev-only mock data) into
a dedicated `__data__/mockConformaResults.ts` file. This reduces the
hook file size and separates test/dev fixtures from production logic.
